requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
html5lib==1.1
idna==3.4
install==1.3.5
pytz==2022.7
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
webencodings==0.5.1
